refactoring: changed folder names according to kebab-case style
