+++ src/components/App/App.jsx
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from "react";
-import AppHeader from "../app-header/AppHeader";
-import BurgerIngredients from "../burger-ingredients/BurgerIngredients";
-import BurgerConstructor from "../burger-constructor/BurgerConstructor";
+import AppHeader from "../app-header/app-header";
+import BurgerIngredients from "../burger-ingredients/burger-ingredients";
+import BurgerConstructor from "../burger-constructor/burger-constructor";
 import appStyle from "./app.module.css";
 import { getIngredients } from "../../utils/api/api";
