+++ src/components/App/App.module.css
@@ -1,7 +1,6 @@
 .app {
   display: flex;
   flex-direction: column;
-
   width: 100%;
   margin: 0 auto;
 }
@@ -10,5 +9,5 @@
   display: flex;
   width: 100%;
   margin: 0 auto;
-  max-width: 1280px;
+  max-width: 1240px;
 }
