+++ src/components/Modal/Modal.jsx
@@ -2,7 +2,7 @@ import { useEffect } from 'react';
 import { createPortal } from 'react-dom';
 import PropTypes from 'prop-types';
 import styles from "./modal.module.css";
-import ModalOverlay from '../modal-overlay/ModalOverlay';
+import ModalOverlay from '../modal-overlay/modal-overlay';
 import { CloseIcon } from '@ya.praktikum/react-developer-burger-ui-components';
 
 const modalContainer = document.getElementById('modal');
