+++ src/components/app-header/app-header.module.css
@@ -13,7 +13,7 @@
   align-items: center;
   margin: 0 auto;
   width: 100%;
-  max-width: 1280px;
+  max-width: 1240px;
 }
 
 .listNavigation {
