+++ src/components/modal-overlay/modal-overlay.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import styles from "./modalOverlay.module.css";
+import styles from "./modal-overlay.module.css";
 import PropTypes from "prop-types";
 
 const ModalOverlay = ({onClose}) => {
